+++ modules/ad_entity_dfp/js/default.view.js
@@ -7,21 +7,18 @@
 
   'use strict';
 
-  if (typeof window.googletag === 'undefined') {
-    window.googletag = {};
-  }
-  var googletag = window.googletag;
-  googletag.cmd = googletag.cmd || [];
-
-  var $window = $(window);
+  window.googletag = window.googletag || {};
+  window.googletag.cmd = window.googletag.cmd || [];
 
   Drupal.ad_entity = Drupal.ad_entity || {};
 
   Drupal.ad_entity.viewHandlers = Drupal.ad_entity.viewHandlers || {};
 
+  var $window = $(window);
+
   Drupal.ad_entity.viewHandlers.dfp_default = Drupal.ad_entity.viewHandlers.dfp_default || {
     initialize: function (containers, context, settings) {
-      googletag.cmd.push(function () {
+      window.googletag.cmd.push(function () {
         var onPageLoad = 'true';
         if (this.numberOfAds > 0) {
           onPageLoad = 'false';
@@ -50,14 +47,14 @@
 
       var slot;
       if (out_of_page === true) {
-        slot = googletag.defineOutOfPageSlot('/' + network_id + '/' + unit_id, ad_id);
+        slot = window.googletag.defineOutOfPageSlot('/' + network_id + '/' + unit_id, ad_id);
       }
       else {
         var sizes = ad_tag.data('dfpSizes');
         if (typeof sizes !== 'object') {
           sizes = [];
         }
-        slot = googletag.defineSlot('/' + network_id + '/' + unit_id, sizes, ad_id);
+        slot = window.googletag.defineSlot('/' + network_id + '/' + unit_id, sizes, ad_id);
       }
 
       ad_tag.data('slot', slot);
@@ -77,34 +74,76 @@
       slot.setTargeting('slotNumber', slotNumber);
       slot.setTargeting('onPageLoad', onPageLoad);
 
-      slot.addService(googletag.pubads());
+      slot.addService(window.googletag.pubads());
     },
     display: function (ad_tags) {
-      var slots = [];
-      for (var i in ad_tags) {
-        if (ad_tags.hasOwnProperty(i)) {
-          var ad_tag = ad_tags[i];
-          var slot = ad_tag.data('slot');
-          if (typeof slot === 'object') {
-            googletag.display(ad_tag.data('id'));
-            slots.push(slot);
+      // When possible, load multiple slots at once to support roadblocks.
+      // Slots with the same ad unit path wouldn't be refreshed
+      // more than one time though, thus they're being split up.
+      var slots = [[]];
+      var slots_length;
+      var slots_list;
+      var i;
+
+      var ad_tags_length = ad_tags.length;
+      for (i = 0; i < ad_tags_length; i++) {
+        var ad_tag = ad_tags[i];
+        var slot = ad_tag.data('slot');
+
+        if (typeof slot === 'object') {
+          window.googletag.display(ad_tag.data('id'));
+
+          var unit_path = slot.getAdUnitPath();
+          slots_length = slots.length;
+          for (var j = 0; j < slots_length; j++) {
+            var exists = false;
+            slots_list = slots[j];
+            var slots_list_length = slots_list.length;
+            for (var k = 0; k < slots_list_length; k++) {
+              if (unit_path === slots_list[k].getAdUnitPath()) {
+                exists = true;
+                break;
+              }
+            }
+            if (exists === false) {
+              slots_list.push(slot);
+              break;
+            }
+            if ((j + 1) === slots.length) {
+              slots.push([slot]);
+            }
           }
         }
       }
-      if (slots.length > 0) {
-        googletag.pubads().refresh(slots);
+
+      slots_length = slots.length;
+      for (i = 0; i < slots_length; i++) {
+        slots_list = slots[i];
+        if (slots_list.length > 0) {
+          window.googletag.pubads().refresh(slots_list);
+        }
       }
     },
     addEventsFor: function (ad_tag, container) {
       // Mark container as initialized once advertisement has been loaded.
-      googletag.pubads().addEventListener('slotRenderEnded', function (event) {
+      var initHandler = function (event) {
         if (event.slot.getSlotElementId() === ad_tag.data('id')) {
           container.removeClass('not-initialized');
           container.addClass('initialized');
           container.data('initialized', true);
+          if (event.isEmpty === true) {
+            container.addClass('empty');
+            container.removeClass('not-empty');
+          }
+          else {
+            container.addClass('not-empty');
+            container.removeClass('empty');
+          }
           container.trigger('adEntity:initialized', [ad_tag, container]);
         }
-      }, false);
+      };
+
+      window.googletag.pubads().addEventListener('slotRenderEnded', initHandler, false);
     },
     numberOfAds: 0
   };