+++ modules/ad_entity_dfp/js/init-provider.js
@@ -8,7 +8,7 @@ googletag.cmd = googletag.cmd || [];
 
 googletag.cmd.push(function () {
   'use strict';
-  googletag.pubads().enableSingleRequest();
+  googletag.pubads().enableSingleRequest(true);
   googletag.pubads().disableInitialLoad();
   googletag.pubads().collapseEmptyDivs();
   googletag.enableServices();